requirements/production.txt
django-cors-headers==4.3.1
django-ratelimit==4.1.0
python-dotenv==1.0.0
whitenoise==6.6.0
gunicorn==21.2.0
django-environ==0.11.2
django-storages==1.14.2
